refactor: extract board data-readers module
Temporary deviation from ARCH-board-server-split.md import graph: getTasks,
readVerdicts, readSecStats still live in server.mjs at this step (they move
to lib/beads.mjs and lib/verdicts.mjs in later steps per the doc's own
extraction order). data-readers.mjs imports them back from ../server.mjs
via a circular ESM import, which is safe because all three are only called
inside function bodies here, never at module-evaluation time. The three
functions are temporarily marked export in server.mjs with a TODO-style
comment; those exports are removed once beads.mjs/verdicts.mjs take over.
